[FIRRTL] Fix LowerTypes for non-NLA hierpaths
Fix a bug in LowerTypes where it would only update hierarchical paths
that are involved in a non-local annotation (NLAs).  This is done in
preparation for more widespread use of hierarchical paths (e.g., by
sv::VerbatimOp) and in preparation for enabling hierarchical paths to
end at the module and not on the final operation.

It follows that this changes the lowering for a hierarchical path
targeting a component/port to generate one new hierarchical path _for
every_ ground type regardless of whether or not the hierarchical path is
used by a non-local annotation.  Previously, new hierarchical paths
would only be generated when used by non-local annotations.

This works by simplifying the way that annotations are updated in
LowerTypes.  Previously, this would update hierarchical paths when
examining annotations and keeping track of an "old symbol" to "new
symbol" mapping.  This commit changes the tracking to track "old
hw::InnerRefAttr" to "new Vector<AnnoTarget>".  This more closely
matches the lowering procedure (as LowerTypes is fundamentally a mapping
of one aggregate type to one or more ground types and results in a nice
simplification of the existing logic.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>