[FIRRTL] Make droppable_name default
This commit changes a default value of namekind from
interesting name to droppable name.

This is more comfortable in the situation where we want
to create a dummy unnamed wire for example
`builder.create<WireOp>(result.type())`. Previously this dummy wire
had an interesting name because default value is interesting.
This commit also changes IR reprsentation.